1d-mini-project/requirements.txt
RPi.GPIO
imutils
numpy
opencv-python
kivy
libdw
pythymiodw
